Make monster matching case insensitive.
This bug was not quite so noticeable because we used fuzzy string matching by
utilizing Levenshtein distance. However, it may cause otherwise correctly
written monster names to be rejected because they were not written in the same
case as the actual monster name.

I didn't actually test if there are any such monsters where it would fail but
at least now we are sure it works correctly.