Update the exporter script.
This new script automatically detects which variant it is being compiled in (of
the variants it knows about). The way monster data is exported from the
resulting Haskell modules is changed so this script is not compatible with the
rest of the Pinobot code at this commit. But in next commit everything is fine!