Switch out IRC library to irc-client, which supports TLS.
This logs activity that Pinobot sees to stdout.

I tested on irc.libera.chat and it connected through TLS no problemo.

I also tested private messaging Pinobot and having Pinobot on a channel
and checked that responses work.

This increases the size of Pinobot dependencies quite a lot, as the new
IRC library uses lenses, and lenses are a big big depency.