Update EvilHack monsters, update extractor to fill in 'mnum' in dummymonst.
The mnum is used in macro r_data(mon) like #define r_data(mon) (&mons[has_erac(mon) ? ERAC(mon)->rmnum : (mon)->mnum])

The extractor had mnum be 0 which messed this macro up. The macro is
used in "hates_material" function which in turn is used by extractor to
check if a monster is silver-hating. So this makes "hates silver" show
up correctly in PinoBot for this variant.