Support compressed, .yaml.gz files. Makes Slex ~20M -> 2M.
Slex variant has unusually large number of monsters compared to all the
other variants.

As of writing this commit, slex is also disabled by default (unlikely
it will be queried through Pinobot), but I wanted to preserve the .yaml
file as a tangible file within the repository (and not just as an
artifact in git commit history).

Makes the source code checkout smaller.

I also tested zstd first, and it actually made a bigger file than .gz. I
chose .gz instead of modern .zst compression, because the zlib library
in Haskell looks much more mature and unlikely to bit-rot.

Pinobot automatically first tries to load .yaml file, and then, if it
does not exist, it will try to load .yaml.gz.

I kept other variant .yaml files uncompressed, so that they are easier
to browse (in GitHub or in other UIs).