Fix use of uninitialized data in extractor.h (prevents SIGSEGV on startup)
In commit 37eef85f932128de11f1f2c4eca3e8cf8d966dc8 I had simplified the
extractor code to not attempt variant auto-detection anymore, because
the auto-detection barely worked anyway and was just not helping me.

But I accidentally also removed code that initialized this
'command_prefix' C string, and code tried to fprintf() that to the
resulting .yaml file. Because it was uninitialized, it crashed. This
commit initializes it to a placeholder value.

Will likely simplify this code further down the line, but as of writing
this commit, I'm focused on updating some variants so that'll be for
later probably.