Update Dnethack and Notnotdnethack.
I had pulled an old version for Notnotdnethack just recently, and it did
not occur to me that it might be old.

Dnethack also had bunch of updates, and I had not updated it properly.

I pulled the branches in the games that looked like they had most recent
development, and updated all.

Neither variant introduced new attacks or technical things I had to
modify any Pinobot or extractor.h code, so this commit is just updating
the .yaml files. Noice.